+++ android/src/main/java/com/reactnative/googlefit/ActivityHistory.java
@@ -11,7 +11,7 @@
 
 package com.reactnative.googlefit;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.text.TextUtils;
 import android.util.Log;