+++ android/src/main/java/com/reactnative/googlefit/GoogleFitManager.java
@@ -12,12 +12,13 @@ package com.reactnative.googlefit;
 
 import android.app.Activity;
 import android.app.Dialog;
+import android.content.Context;
 import android.content.Intent;
 import android.content.IntentSender;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.util.Log;
 import java.util.ArrayList;
 
@@ -35,7 +36,6 @@ import com.google.android.gms.common.api.ResultCallback;
 import com.google.android.gms.common.api.Scope;
 import com.google.android.gms.common.api.Status;
 import com.google.android.gms.fitness.Fitness;
-import com.google.android.gms.fitness.Fitness.*;
 import com.google.android.gms.auth.api.signin.*;
 
 
@@ -54,10 +54,12 @@ public class GoogleFitManager implements
     private BodyHistory bodyHistory;
     private HeartrateHistory heartrateHistory;
     private CalorieHistory calorieHistory;
+    private NutritionHistory nutritionHistory;
     private StepCounter mStepCounter;
     private StepSensor stepSensor;
     private RecordingApi recordingApi;
     private ActivityHistory activityHistory;
+    private HydrationHistory hydrationHistory;
 
     private static final String TAG = "RNGoogleFit";
 
@@ -75,8 +77,10 @@ public class GoogleFitManager implements
         this.heartrateHistory = new HeartrateHistory(mReactContext, this);
         this.distanceHistory = new DistanceHistory(mReactContext, this);
         this.calorieHistory = new CalorieHistory(mReactContext, this);
+        this.nutritionHistory = new NutritionHistory(mReactContext, this);
         this.recordingApi = new RecordingApi(mReactContext, this);
         this.activityHistory = new ActivityHistory(mReactContext, this);
+        this.hydrationHistory = new HydrationHistory(mReactContext, this);
         //        this.stepSensor = new StepSensor(mReactContext, activity);
     }
 
@@ -117,6 +121,10 @@ public class GoogleFitManager implements
 
     public CalorieHistory getCalorieHistory() { return calorieHistory; }
 
+    public NutritionHistory getNutritionHistory() { return nutritionHistory; }
+
+    public HydrationHistory getHydrationHistory() { return hydrationHistory; }
+
     public void authorize(ArrayList<String> userScopes) {
         final ReactContext mReactContext = this.mReactContext;
 
@@ -131,57 +139,64 @@ public class GoogleFitManager implements
 
         mApiClient = apiClientBuilder
                 .addConnectionCallbacks(
-                    new GoogleApiClient.ConnectionCallbacks() {
-                        @Override
-                        public void onConnected(@Nullable Bundle bundle) {
-                            Log.i(TAG, "Authorization - Connected");
-                            sendEvent(mReactContext, "GoogleFitAuthorizeSuccess", null);
-                        }
+                        new GoogleApiClient.ConnectionCallbacks() {
+                            @Override
+                            public void onConnected(@Nullable Bundle bundle) {
+                                Log.i(TAG, "Authorization - Connected");
+                                sendEvent(mReactContext, "GoogleFitAuthorizeSuccess", null);
+                            }
 
-                        @Override
-                        public void onConnectionSuspended(int i) {
-                            Log.i(TAG, "Authorization - Connection Suspended");
-                            if ((mApiClient != null) && (mApiClient.isConnected())) {
-                                mApiClient.disconnect();
+                            @Override
+                            public void onConnectionSuspended(int i) {
+                                Log.i(TAG, "Authorization - Connection Suspended");
+                                if ((mApiClient != null) && (mApiClient.isConnected())) {
+                                    mApiClient.disconnect();
+                                }
                             }
                         }
-                    }
                 )
                 .addOnConnectionFailedListener(
-                    new GoogleApiClient.OnConnectionFailedListener() {
-                        @Override
-                        public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
-                            Log.i(TAG, "Authorization - Failed Authorization Mgr:" + connectionResult);
-                            if (mAuthInProgress) {
-                                Log.i(TAG, "Authorization - Already attempting to resolve an error.");
-                            } else if (connectionResult.hasResolution()) {
-                                try {
+                        new GoogleApiClient.OnConnectionFailedListener() {
+                            @Override
+                            public void onConnectionFailed(@NonNull ConnectionResult connectionResult) {
+                                Log.i(TAG, "Authorization - Failed Authorization Mgr:" + connectionResult);
+                                if (mAuthInProgress) {
+                                    Log.i(TAG, "Authorization - Already attempting to resolve an error.");
+                                } else if (connectionResult.hasResolution()) {
+                                    try {
+                                        mAuthInProgress = true;
+                                        connectionResult.startResolutionForResult(mActivity, REQUEST_OAUTH);
+                                    } catch (IntentSender.SendIntentException e) {
+                                        Log.i(TAG, "Authorization - Failed again: " + e);
+                                        mApiClient.connect();
+                                    }
+                                } else {
+                                    Log.i(TAG, "Show dialog using GoogleApiAvailability.getErrorDialog()");
+                                    showErrorDialog(connectionResult.getErrorCode());
                                     mAuthInProgress = true;
-                                    connectionResult.startResolutionForResult(mActivity, REQUEST_OAUTH);
-                                } catch (IntentSender.SendIntentException e) {
-                                    Log.i(TAG, "Authorization - Failed again: " + e);
-                                    mApiClient.connect();
+                                    WritableMap map = Arguments.createMap();
+                                    map.putString("message", "" + connectionResult);
+                                    sendEvent(mReactContext, "GoogleFitAuthorizeFailure", map);
                                 }
-                            } else {
-                                Log.i(TAG, "Show dialog using GoogleApiAvailability.getErrorDialog()");
-                                showErrorDialog(connectionResult.getErrorCode());
-                                mAuthInProgress = true;
-                                WritableMap map = Arguments.createMap();
-                                map.putString("message", "" + connectionResult);
-                                sendEvent(mReactContext, "GoogleFitAuthorizeFailure", map);
                             }
                         }
-                    }
                 )
                 .build();
 
         mApiClient.connect();
     }
 
-    public void  disconnect() {
+    public void  disconnect(Context context) {
+        GoogleSignInOptions options = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestEmail()
+                .build();
+        GoogleSignInClient googleSignInClient = GoogleSignIn.getClient(context, options);
+
         GoogleSignInAccount gsa = GoogleSignIn.getAccountForScopes(mReactContext, new Scope(Scopes.FITNESS_ACTIVITY_READ));
         Fitness.getConfigClient(mReactContext, gsa).disableFit();
         mApiClient.disconnect();
+
+        googleSignInClient.signOut();
     }
 
     public boolean isAuthorized() {