+++ android/src/main/java/com/reactnative/googlefit/GoogleFitModule.java
@@ -26,7 +26,6 @@ import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;
 import com.facebook.react.bridge.LifecycleEventListener;
-import com.facebook.react.bridge.WritableArray;
 import com.facebook.react.uimanager.IllegalViewOperationException;
 import com.google.android.gms.fitness.data.DataType;
 import com.google.android.gms.fitness.data.HealthDataTypes;
@@ -108,9 +107,14 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
     }
 
     @ReactMethod
-    public void disconnect() {
-        if (mGoogleFitManager != null) {
-            mGoogleFitManager.disconnect();
+    public void disconnect(Promise promise) {
+        try {
+            if (mGoogleFitManager != null) {
+                mGoogleFitManager.disconnect(getCurrentActivity());
+            }
+            promise.resolve(null);
+        } catch (Exception e) {
+            promise.reject(e);
         }
     }
 
@@ -147,6 +151,34 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
         }
     }
 
+     @ReactMethod
+    public void getDailyHalfHourData(double startDate,
+                                         double endDate,
+                                         Callback errorCallback,
+                                         Callback successCallback) {
+
+        try {
+          
+            successCallback.invoke(mGoogleFitManager.getStepHistory().aggregateDataByHalfHour((long) startDate, (long) endDate));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
+    @ReactMethod
+    public void getDailyHourData(double startDate,
+                                     double endDate,
+                                     Callback errorCallback,
+                                     Callback successCallback) {
+
+        try {
+           
+            successCallback.invoke(mGoogleFitManager.getStepHistory().aggregateDataByHour((long) startDate, (long) endDate));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
     @ReactMethod
     public void getActivitySamples(double startDate,
                                    double endDate,
@@ -155,6 +187,19 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
 
         try {
             successCallback.invoke(mGoogleFitManager.getActivityHistory().getActivitySamples((long)startDate, (long)endDate));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
+    @ReactMethod
+    public void getUserInputSteps(double startDate,
+                                double endDate,
+                                Callback errorCallback,
+                                Callback successCallback) {
+
+        try {
+            mGoogleFitManager.getStepHistory().getUserInputSteps((long) startDate, (long) endDate, successCallback);
         } catch (IllegalViewOperationException e) {
             errorCallback.invoke(e.getMessage());
         }
@@ -238,6 +283,18 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
                          Callback successCallback) {
         try {
             successCallback.invoke(mGoogleFitManager.getCalorieHistory().saveFood(foodSample));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
+    @ReactMethod
+    public void getDailyNutritionSamples(double startDate,
+                                         double endDate,
+                                         Callback errorCallback,
+                                         Callback successCallback) {
+        try {
+            successCallback.invoke(mGoogleFitManager.getNutritionHistory().aggregateDataByDate((long) startDate, (long) endDate));
         } catch (IllegalViewOperationException e) {
             errorCallback.invoke(e.getMessage());
         }
@@ -328,9 +385,9 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
 
     @ReactMethod
     public void getBloodPressureSamples(double startDate,
-                                 double endDate,
-                                 Callback errorCallback,
-                                 Callback successCallback) {
+                                        double endDate,
+                                        Callback errorCallback,
+                                        Callback successCallback) {
         try {
             HeartrateHistory heartrateHistory = mGoogleFitManager.getHeartrateHistory();
             heartrateHistory.setDataType(HealthDataTypes.TYPE_BLOOD_PRESSURE);
@@ -342,14 +399,47 @@ public class GoogleFitModule extends ReactContextBaseJavaModule implements Lifec
 
     @ReactMethod
     public void getHeartRateSamples(double startDate,
-                                 double endDate,
-                                 Callback errorCallback,
-                                 Callback successCallback) {
+                                    double endDate,
+                                    Callback errorCallback,
+                                    Callback successCallback) {
 
         try {
             HeartrateHistory heartrateHistory = mGoogleFitManager.getHeartrateHistory();
             heartrateHistory.setDataType(DataType.TYPE_HEART_RATE_BPM);
             successCallback.invoke(heartrateHistory.getHistory((long)startDate, (long)endDate));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
+    @ReactMethod
+    public void getHydrationSamples(double startDate,
+                                    double endDate,
+                                    Callback errorCallback,
+                                    Callback successCallback) {
+        try {
+            successCallback.invoke(mGoogleFitManager.getHydrationHistory().getHistory((long) startDate, (long) endDate));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+
+    @ReactMethod
+    public void saveHydration(ReadableArray hydrationArray,
+                           Callback errorCallback,
+                           Callback successCallback) {
+        try {
+            HydrationHistory hydrationHistory = mGoogleFitManager.getHydrationHistory();
+            successCallback.invoke(hydrationHistory.save(hydrationArray));
+        } catch (IllegalViewOperationException e) {
+            errorCallback.invoke(e.getMessage());
+        }
+    }
+    @ReactMethod
+    public void deleteHydration(ReadableMap options, Callback errorCallback, Callback successCallback) {
+        try {
+            HydrationHistory hydrationHistory = mGoogleFitManager.getHydrationHistory();
+            successCallback.invoke(hydrationHistory.delete(options));
         } catch (IllegalViewOperationException e) {
             errorCallback.invoke(e.getMessage());
         }